move merge functions into merge module
